traffic_ingestor/requirements.txt
azure-functions
python-dotenv
requests
azure-data-tables
azure-eventgrid
azure-core
